Update to 1.0.3.0
* Crafted items now correctly have their defaults set.
* Item and tooltip renaming is now less hacky and conforms with how
Terraria handles localization. (You'll still need to restart the game in
order to remove changes made to names or tooltips)
* Now tells you if your executable has already been patched.
* Required files are now copied first so that if an error occurs at that
point, the executable will still remain untouched.
* Every required file and dll is now embedded into the executable.
* Now waits half a second before making the executable large address
aware. This should fix an issue where it couldn't write to the exe
because it wasn't closed by the patching library yet.
* Now asks if you'd like to cleanup required files from the executable
directory on restore.
* A bit of code refactoring.